Update client framework

+++ StayWell.ServiceDefinitions/Content/Objects/ContentRatingResponse.cs
@@ -7,7 +7,7 @@ namespace StayWell.ServiceDefinitions.Content.Objects
         public Guid ContentId { get; set; }
         public ContentType ContentType { get; set; }
         public Guid LicenseId { get; set; }
-        public decimal Rating { get; set; }
+        public double Rating { get; set; }
         public long RatingsCount { get; set; }
         public long ViewCount { get; set; }
     }
